Update Privacy.mdx
Hi Jared - Ciera asked me to help with transferring an SOP from an external document into GitHub. I was able to figure out most of what I needed to do, except 1) how to change font size, 2) how to change font color, and 3) how to create a table. Therefore this draft is not complete - but I wanted to be sure to save the changes I've made so far. 
I'm hoping that creating a new branch is the correct next step. Thanks! -marty